requirements.txt
openai==1.99.6
python-dotenv
openai-agents